Add Linux support
All examples work excellently! Tested against NDI SDK version 6.3.2.0

Major adjustments to the existing library:
- All enums which are directly based on C enums use repr(u32) on Linux, since enums are `unsigned int` on Linux (at least bindgen assumes so) and `int` on Windows. This is not a breaking change, since the library didn’t compile on Linux beforehand, though users now cannot rely on the repr(X) attribute.
- The absence of `unsafe extern "C"` in the stub bindings is no longer enforced, since the Linux bindings contain C function pointers which are annotated as such. Since the stubs are not regenerated on docs.rs anyways, this should not break anything.

Linux build process: We assume libndi.so is in a default library path, which should be the case if the user installed the library in any reasonable way. (Otherwise, they can set LDPATH as normal.) For the headers, we assume it’s in the standard /usr/include and require setting NDI_HEADER_DIR otherwise (very friendly error message provided).

Added a neat platform support table in case someone wants to add MacOS :>

+++ src/frame/video.rs
@@ -238,7 +238,7 @@ impl VideoFrame {
     }
 
     pub fn raw_four_cc(&self) -> FourCC {
-        FourCC::from_ffi(self.raw.FourCC)
+        FourCC::from_ffi(self.raw.FourCC as i32)
     }
 
     /// Sets the FourCC format of the frame.
